relocate registry
