+++ deepecohab/utils/auxfun.py
@@ -3,60 +3,13 @@ from itertools import product
 from pathlib import Path
 from typing import (
 	Any,
-	Callable,
 	Literal,
-	Iterator,
 )
 
 import polars as pl
 import toml
 
-
-class DataFrameRegistry:
-	def __init__(self):
-		self._registry: dict[str, Callable] = {}
-		self.analysis_steps: list[str] = [
-			"activity_df",
-			"cage_occupancy",
-			"chasings_df",
-			"tube_test_df",
-			"ranking",
-			"pairwise_meetings",
-			"incohort_sociability",
-			"time_alone",
-			"pairwise_meetings",
-			"feature_df",
-		]
-
-	def register(self, name: str):
-		"""Decorator to register a new plot type."""
-
-		def wrapper(func: Callable):
-			self._registry[name] = func
-			return func
-
-		return wrapper
-
-	def list_available(self) -> list[str]:
-		"""Returns a list of all registered function names."""
-		return list(self._registry.keys())
-
-	def run_pipeline(
-		self, config: dict[str, Any], **kwargs
-	) -> Iterator[tuple[str, int, list[int]]]:
-		"""Runs the pipeline and yields status updates.
-
-		Yields:
-			(step_name, current_index, total_steps)
-		"""
-		total = len(self.analysis_steps)
-		for i, name in enumerate(self.analysis_steps):
-			yield name, i + 1, total
-			func = self._registry[name]
-			func(config, **kwargs)
-
-
-df_registry = DataFrameRegistry()
+from deepecohab.core.registries import df_registry
 
 
 def read_config(config_path: str | Path | dict[str, Any]) -> dict:
