+++ deepecohab/utils/cache_config.py
@@ -3,6 +3,7 @@ from pathlib import Path
 import diskcache
 from dash import DiskcacheManager
 
+from deepecohab.core.registries import df_registry
 from deepecohab.utils import auxfun
 
 cache_dir = Path(r"\cache")
@@ -22,6 +23,6 @@ def get_project_data(config_tuple):
 
 	return {
 		name: auxfun.load_ecohab_data(config, name, return_df=True)
-		for name in auxfun.df_registry.list_available()
+		for name in df_registry.list_available()
 		if "binary" not in str(name)
 	}
